feat: expose WithSlotProps type
