+++ packages/core/types/Internal.tsx
@@ -28,14 +28,3 @@ export type PrivateAppState<UserData extends Data = Data> =
       zones: ZoneIndex;
     };
   };
-
-export type WithSlotProps<
-  Props extends DefaultComponentProps = DefaultComponentProps,
-  SlotProps extends DefaultComponentProps = Props
-> = Props extends any
-  ? any
-  : {
-      [PropName in keyof Props]: Props[PropName] extends Slot<SlotProps>
-        ? Content<SlotProps>
-        : Props[PropName];
-    };
